Move violation logic to a private method
